Add files via upload
Plan is to work out the edge graph and rates via cpu before moving into compute buffer.
So far, I am not sure I am getting the mesh info correctly, or updating it to the renderer correctly.
Some parts seem to update, some not, and connectivity of the graph seldom makes sense. So most of this code is trying to visualize exactly how the mesh is connected.